+++ src/main/webapp/src/pages/Blog.js
@@ -1,8 +1,8 @@
 import React from "react";
-import "../../styles/blog.scss";
-import BLOGRest from "../../scripts/blog-rest";
-import BlogPost from "./Blogpost";
-import Banner from "../Header";
+import "../styles/blog.scss";
+import BLOGRest from "../scripts/blog-rest";
+import BlogPost from "../components/blog/Blogpost";
+import Banner from "../components/Header";
 import {
   Avatar,
   BorderBox,
@@ -14,8 +14,9 @@ import {
   Text
 } from "@primer/components";
 import {theme} from "@primer/components/lib/theme-preval";
-import {LAYOUT_WIDTH} from "../../Constants";
+import {LAYOUT_WIDTH} from "../Constants";
 import "@primer/css/utilities/index.scss";
+import Footer from "../components/Footer";
 
 export default class Blog extends React.Component {
 
@@ -42,24 +43,25 @@ export default class Blog extends React.Component {
     return (
         <>
           <Banner/>
-          <BorderBox boxShadow={theme.shadows.medium} backgroundColor="white"
+          <BorderBox boxShadow={theme.shadows.medium} backgroundColor="canvas"
                      maxWidth={LAYOUT_WIDTH} mx="auto" my={3} p={3}
                      as="main">
-            <Pagehead flexGrow={1} px={3}>
+            <Pagehead flexGrow={1}>
               <Heading as="h1">
                 Mathias Bosman // Blog
               </Heading>
             </Pagehead>
 
             <section>
-              <Flex as="article">
+              <Flex as="article" className="flexWrapper" my={4}>
                 <Box flexGrow={1}>
                   <Link href="#" className="text-mono" color="gray.5" muted>
                     <time dateTime="2021-02-25">February 25, 2021</time>
                   </Link>
                 </Box>
                 <Box flexGrow={3}>
-                  <Flex as="ul" mb={3} className="list-style-none text-mono categories">
+                  <Flex as="ul" mb={3}
+                        className="list-style-none text-mono categories">
                     <li><Link muted href="#">category 1</Link></li>
                     <li><Link muted href="#">category 2</Link></li>
                   </Flex>
@@ -67,10 +69,11 @@ export default class Blog extends React.Component {
                   <Box>
                     <p>Here's some text bro! Change this to paragraph</p>
                   </Box>
-                  <Link href="#" muted className="author">
+                  <Link href="#" className="author" fontSize={1} color="gray.6">
                     <Flex alignItems="center">
-                    <Avatar src="/assets/images/avatar.webp" size={35} mr={3}/>
-                    <Text>Mathias B.</Text>
+                      <Avatar src="/assets/images/avatar.webp" size={35}
+                              mr={3}/>
+                      <Text>Mathias B.</Text>
                     </Flex>
                   </Link>
                 </Box>
@@ -80,6 +83,7 @@ export default class Blog extends React.Component {
             <section>
               {this._renderPosts()}
             </section>
+            <Footer/>
           </BorderBox>
         </>
     );
